backend/requirements.txt
# Django & DRF
Django==4.2.16
djangorestframework==3.15.2
django-cors-headers==4.4.0
djangorestframework-simplejwt==5.3.1
django-filter==24.3
drf-yasg==1.21.7

# Database
psycopg2-binary==2.9.9

# Image / File Processing
Pillow==10.4.0
opencv-python==4.10.0.84
pytesseract==0.3.13
qrcode==7.4.2
WeasyPrint==62.3

# Rich Text / Admin
django-ckeditor==6.7.2

# Task Queue & Cache
celery==5.4.0
redis==5.1.1

# History / Audit
django-simple-history==3.7.0

# Utilities
openpyxl==3.1.5
python-decouple==3.8
python-dotenv==1.0.1
gunicorn==23.0.0
whitenoise==6.7.0
django-crispy-forms==2.3
crispy-bootstrap5==2024.10
